Added optional force termination (#11)

+++ src/hotpocket-contract.js
@@ -12,8 +12,9 @@ export class HotPocketContract {
 
     #controlChannel = null;
     #clientProtocol = null;
+    #forceTerminate = false;
 
-    init(contractFunc, clientProtocol = clientProtocols.json) {
+    init(contractFunc, clientProtocol = clientProtocols.json, forceTerminate = false) {
 
         return new Promise(resolve => {
             if (this.#controlChannel) { // Already initialized.
@@ -29,6 +30,8 @@ export class HotPocketContract {
                 resolve(false);
                 return;
             }
+
+            this.#forceTerminate = forceTerminate;
 
             // Parse HotPocket args.
             fs.readFile(process.stdin.fd, 'utf8', (err, argsJson) => {
@@ -60,6 +63,7 @@ export class HotPocketContract {
 
     #terminate() {
         this.#controlChannel.close();
-        process.kill(process.pid, 'SIGINT');
+        if (this.#forceTerminate)
+            process.kill(process.pid, 'SIGINT');
     }
 }
